% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{breastcancer_data_subset}
\alias{breastcancer_data_subset}
\title{First five letters of the alphabet}
\format{
A data frame with 5 rows and 2 variables:
\describe{
\item{patient_id}{integer from 1 to 5}
\item{image_number}{letter, from A to E}
\item{cell_id}{letter, from A to E}
\item{cell_x}{letter, from A to E}
\item{cell_y}{letter, from A to E}
\item{cell_metacluster}{letter, from A to E}
\item{patient_age}{letter, from A to E}
\item{tumor_grade}{letter, from A to E}
}
}
\source{
\url{https://bodenmillergroup.github.io/imcdatasets/reference/JacksonFischer_2020_BreastCancer.html}
}
\usage{
breastcancer_data_subset
}
\description{
A dataset containing first five letters of English alphabet
}
\keyword{datasets}
