% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_comb_CB.R
\name{calc.preds}
\alias{calc.preds}
\title{Preprocesses data}
\usage{
calc.preds(beta_hat, lin.comb)
}
\arguments{
\item{grid}{the grid that the CBs were evaluated over.  eg. 0:200}
}
\value{
It returns res.wildBS
}
\description{
Preprocesses data
}
