% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_package_SEs}
\alias{extract_package_SEs}
\title{Extracts package SE's}
\usage{
extract_package_SEs(name, model, grid)
}
\arguments{
\item{name}{A \emph{pffr()} model object}

\item{model}{A \emph{pffr()} model object}

\item{grid}{The grid on which the spatial summary function is evaluated.  For example, 0:200}
}
\value{
A vector containing the target curve.
}
\description{
Extracts package SE's
}
