% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_target_curve}
\alias{extract_target_curve}
\title{Extracts a target curve from a \code{refund::pffr()} model object}
\usage{
extract_target_curve(model, grid, target, form, covar_list = list(), func)
}
\arguments{
\item{model}{A fitted model object obtained from the output of \code{refund::pffr()}}

\item{grid}{The grid of evaluation radii on which the spatial summary function is calculated.  For example, \code{0:200}}

\item{target}{A character vector containing the names of the model terms to be combined to form the target curve for which to construct the confidence band around.  For example, \code{c("Intercept", "var1", "var2")}}

\item{form}{A character vector of equal length as \code{target}, with each entry corresponding to an entry in \code{target}.  Each entry must be either \code{"coef"} or \code{"term"}, where \code{"coef"} uses the estimated coefficient function directly and \code{"term"} multiplies the coefficient function by the corresponding value in \code{covar_list}.}

\item{covar_list}{A named list giving the covariate values used when constructing target-curve components with \code{form = "term"}. The names of \code{covar_list} must match \code{target}.  For example, if \code{target = c("Intercept", "var1", "var2")}, then \code{covar_list} could be \code{list(Intercept = rep(1, length(v1)), var1 = v1, var2 = v2)}, where \code{v1} and \code{v2} each have the length of the grid of evaluation radii.}

\item{func}{A function specifying how the terms in \code{target} are combined to form the target curve. The argument names of \code{func} must match \code{target}. For example, if \code{target = c("Intercept", "var1", "var2")}, then \code{func} might be \code{function(Intercept, var1, var2){Intercept + var1 + var2}}.}
}
\value{
A vector containing estimated values of the target curve for each evaluation radius.
}
\description{
Extracts a target curve from a \code{refund::pffr()} model object
}
