% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EY_CB.R
\name{EY_CB}
\alias{EY_CB}
\title{Calculates confidence bands for \eqn{\mathbb{E}[Y_i(r)]} using \emph{pffr()} estimates of model coefficient functions and two layers of wild bootstrap resampling}
\usage{
EY_CB(
  formula,
  covar.list = list(),
  data,
  spatial.covars = NULL,
  B1 = 500,
  B2 = 100,
  alpha = 0.05,
  re = NULL,
  n_cores = NULL,
  seed = NULL
)
}
\arguments{
\item{formula}{An object of class \emph{formula}.  The formula for the \emph{pffr()} model with special terms as in mgcv's GAM}

\item{covar.list}{a list containing the values of the covariates across radii in the linear combination.
For example, if we want a confidence band for \eqn{\beta_0(r) + \beta_1(r)Group + \beta_2(r)L_{AB}(r)} for patient 1, use \strong{covar.list} = list(group = vec1, Lab = vec2), where vec1 and vec2 are R-length vectors containing covariate values for patient 1 at each radii, in ascending order of radii.  Covariates in covar.list should be specified in the order that they appear in the formula statement.  If the formula statement contains random effects, do not include them in covar.list
To calculate a confidence band for \eqn{\beta_0(r)} only, use \strong{covar.list}=list()}

\item{data}{Data frame obtained from the output of \emph{funboot::preprocess_data()}}

\item{spatial.covars}{Vector containing the names of the covariate(s) to be treated as spatially-varying.  For example, c('var1','var2')}

\item{B1}{Number of bootstrap samples in the outside layer}

\item{B2}{Number of bootstrap samples in the inside layer}

\item{alpha}{Desired significance level for the confidence band. For example, 0.05}

\item{re}{Vector containing names of variable(s) for which to fit a random intercept.  For example, c('patient_id'). The variable must also be included in the formula argument, using special terms as in mgcv's GAM}

\item{n_cores}{Number of cores to use during parallel processing.  If NULL, parallel::detectCores(logical=FALSE)-1 cores are used}

\item{seed}{Random seed to be used during the bootstrap resampling (optional).}
}
\value{
A list containing the lower and upper bounds of the confidence band, as well as estimated model coefficients from the \emph{pffr()} output and
additional intermittent variables used in the calculation of the confidence band.
'M1' is the bootstrap distribution of M from the outer layer.  'bs.sd' are the bootstrap standard deviations for each radius, yielded by the inner layer of resampling
}
\description{
Calculates confidence bands for \eqn{\mathbb{E}[Y_i(r)]} using \emph{pffr()} estimates of model coefficient functions and two layers of wild bootstrap resampling
}
