% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_fixed_and_re}
\alias{get_fixed_and_re}
\title{Extracts estimated fixed effects and random effects from a \code{refund::pffr()} model object}
\usage{
get_fixed_and_re(model, n_Im_total, grid, id)
}
\arguments{
\item{model}{A fitted model object obtained from the output of \code{refund::pffr()}}

\item{n_Im_total}{The total number of observed images}

\item{grid}{The grid of evaluation radii on which the spatial summary function is calculated.  For example, \code{0:200}}

\item{id}{The grouping variable whose name defines the random effect in the fitted model.  For example, \code{"patient_id"}}
}
\value{
A list containing estimated fixed effects and random effects at each evaluation radius
}
\description{
Extracts estimated fixed effects and random effects from a \code{refund::pffr()} model object
}
