% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_comb_CB.R
\name{lin_comb_CB}
\alias{lin_comb_CB}
\title{Calculates wild bootstrap confidence bands for linear combinations of \emph{pffr()} model coefficient functions}
\usage{
lin_comb_CB(
  formula,
  lin.comb = list(),
  data,
  spatial.covars = NULL,
  B1 = 500,
  B2 = 100,
  alpha = 0.05,
  re = NULL,
  n_cores = NULL,
  seed = NULL
)
}
\arguments{
\item{formula}{An object of class \emph{formula}.  The formula for the \emph{pffr()} model}

\item{lin.comb}{a list containing the values of the covariates across radii in the linear combination.
For example, if we want a confidence band for the linear combination \eqn{\beta_0(r) + \beta_1(r)Group + \beta_2(r)L_{AB}(r)} for patient 1, use \strong{lin.comb} = list(group = data\link{data$patient_id=='1',}$group, Lab = data\link{data$patient_id=='1',}$Lab).
To calculate a confidence band for \eqn{\beta_0(r)} only, use \strong{lin.comb}=list()}

\item{data}{Data frame obtained from the output of \emph{funboot::preprocess_data()}}

\item{spatial.covars}{Vector containing the names of the covariate(s) to be specified as spatially-varying.  For example, c('var1','var2')}

\item{B1}{Number of bootstrap samples in the outside layer}

\item{B2}{Number of bootstrap samples in the inside layer}

\item{alpha}{Desired significance level for the confidence band. For example, 0.05}

\item{re}{Vector containing names of variable(s) for which to fit a random intercept.  For example, c('patient_id')}

\item{n_cores}{Number of cores to use during parallel processing.  If NULL, parallel::detectCores(logical=FALSE)-1 cores are used}

\item{seed}{Random seed to be used during the bootstrap resampling (optional).}
}
\value{
A list containing the lower and upper bounds of the confidence band, as well as estimated model coefficients from the \emph{pffr()} output and
additional intermittent variables used in the calculation of the confidence band.  'p.vals' are p-values for each model coefficient function yielded by the wild boostrap test of the model coefficient against 0, and 'test.stats' are the associated test statistics.
'M1' is the bootstrap distribution of M from the outer layer.  'bs.SE' are the bootstrap standard errors for each radius, yielded by the inner layer of resampling
}
\description{
Calculates wild bootstrap confidence bands for linear combinations of \emph{pffr()} model coefficient functions
}
