% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melanoma_data.R
\docType{data}
\name{melanoma_data}
\alias{melanoma_data}
\title{Hoch-Schulz Melanoma Data Subset}
\format{
A data frame with 307931 rows and 7 variables:
\describe{
\item{patient_id}{Patient ID}
\item{image_number}{Image Number}
\item{cell_id}{Cell ID}
\item{cell_x}{X coordinate of the cell on the image}
\item{cell_y}{Y coordinate of the cell on the image}
\item{cell_type}{Cell type}
\item{patient_age}{Patient age in years}
}
}
\usage{
melanoma_data
}
\description{
DESCRIPTION
}
\keyword{datasets}
