% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.wildBS_CB.R
\name{plot.wildBS_CB}
\alias{plot.wildBS_CB}
\title{Preprocesses data}
\usage{
\method{plot}{wildBS_CB}(CB.object = CBs, alpha = 0.05)
}
\arguments{
\item{grid}{the grid that the CBs were evaluated over.  eg. 0:200}
}
\value{
It returns res.wildBS
}
\description{
Preprocesses data
}
