% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_EY_CB.R
\name{plot_EY_CB}
\alias{plot_EY_CB}
\title{Plots wild bootstrap confidence bands obtained from \emph{funboot::EY_CB()}}
\usage{
plot_EY_CB(CB.object = CBs)
}
\arguments{
\item{CB.object}{List obtained from the output of \emph{funboot::EY_CB()}}
}
\value{
A \emph{ggplot} object
}
\description{
Plots wild bootstrap confidence bands obtained from \emph{funboot::EY_CB()}
}
