% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lin_comb_CB.R
\name{plot_lin_comb_CB}
\alias{plot_lin_comb_CB}
\title{Plots wild bootstrap confidence bands obtained from \emph{funboot::lin_comb_CB()}}
\usage{
plot_lin_comb_CB(CB.object = CBs)
}
\arguments{
\item{CB.object}{List obtained from the output of \emph{funboot::lin_comb_CB()}}
}
\value{
A \emph{ggplot} object
}
\description{
Plots wild bootstrap confidence bands obtained from \emph{funboot::lin_comb_CB()}
}
