% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{plot_model}
\alias{plot_model}
\title{Preprocesses data}
\usage{
plot_model(CB.object = CBs, alpha = 0.05)
}
\arguments{
\item{grid}{the grid that the CBs were evaluated over.  eg. 0:200}
}
\value{
It returns res.wildBS
}
\description{
Preprocesses data
}
