% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wildBS_CB.R
\name{plot_wildBS_CB}
\alias{plot_wildBS_CB}
\title{Plots target curve estimates and wild bootstrap confidence bands obtained from \code{funboot::wildBS_CB()}}
\usage{
plot_wildBS_CB(CB_object)
}
\arguments{
\item{CB_object}{List obtained from the output of \code{funboot::wildBS_CB()}}
}
\value{
A \code{ggplot} object showing the estimated target curve and its wild bootstrap confidence band
}
\description{
Plots target curve estimates and wild bootstrap confidence bands obtained from \code{funboot::wildBS_CB()}
}
