% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointwise_test.R
\name{pointwise_test}
\alias{pointwise_test}
\title{Preprocesses data}
\usage{
pointwise_test(
  data,
  formula = formula,
  r.star = 100,
  re = c("patient_id"),
  alpha = 0.05,
  spatial.covars = NULL,
  method = c("Bonferroni")
)
}
\arguments{
\item{sumfun.data}{is}
}
\value{
It returns res.wildBS
}
\description{
Preprocesses data
}
