% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_wildBS.R
\name{test_wildBS}
\alias{test_wildBS}
\title{Preprocesses data}
\usage{
test_wildBS(sumfun.data, image.dims, seed = 456, parallel = T)
}
\arguments{
\item{sumfun.data}{is}
}
\value{
It returns res.wildBS
}
\description{
Preprocesses data
}
