% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EY_CB.R
\name{calc.preds}
\alias{calc.preds}
\title{Calculates predicted values using \emph{pffr()} estimates of model coefficient functions}
\usage{
calc.preds(beta_hat, covar.list)
}
\arguments{
\item{beta_hat}{An R x p matrix of values containing the estimated model coefficient functions from \emph{pffr()}}

\item{covar.list}{a list containing the values of the covariates across radii.
For example, if we want a confidence band for \eqn{\beta_0(r) + \beta_1(r)Group + \beta_2(r)L_{AB}(r)} for patient 1, use \strong{covar.list} = list(group = vec1, Lab = vec2), where vec1 and vec2 are R-length vectors containing covariate values for patient 1 at each radii, in ascending order of radii
To calculate a confidence band for \eqn{\beta_0(r)} only, use \strong{covar.list}=list()}
}
\value{
An R-length vector containing the predicted values, in ascending order of radii
}
\description{
Calculates predicted values using \emph{pffr()} estimates of model coefficient functions
}
