% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_package_SEs}
\alias{extract_package_SEs}
\title{Extracts the target curve (e.g. the curve to create the wild bootstrap confidence band around) from a \emph{pffr()} model object}
\usage{
extract_package_SEs(name, model, grid)
}
\arguments{
\item{model}{A \emph{pffr()} model object}

\item{grid}{The grid on which the spatial summary function is evaluated.  For example, 0:200}

\item{target}{A vector containing the names of variables and/or coefficients used to calculate the target curve.  For example, c("var1","var2").  If the target curve includes the intercept curve, specify it as "Intercept".}

\item{form}{A vector of equal length as \emph{target}.  Each element should be be either "coef" or "term", depending on whether the the corresponding element of \emph{target} represents a coefficient curve (e.g. \eqn{\beta_1(r)}) or full term (e.g. \eqn{\beta_1(r)X_1(r)}).}

\item{covar.list}{a list containing the values of the covariates across radii.
For example, if we want a confidence band for \eqn{\beta_0(r) + \beta_1(r)Group + \beta_2(r)L_{AB}(r)} for patient 1, use \strong{covar.list} = list(group = vec1, Lab = vec2), where vec1 and vec2 are R-length vectors containing covariate values for patient 1 at each radii, in ascending order of radii
The names of \emph{covar.list} must match the names of \emph{target}.
To calculate a confidence band for \eqn{\beta_0(r)} only, use \strong{covar.list}=list().}

\item{func}{The function of \emph{form} which produces the target curve.  Its arguments should match \emph{target}.}
}
\value{
A vector containing the target curve.
}
\description{
Extracts the target curve (e.g. the curve to create the wild bootstrap confidence band around) from a \emph{pffr()} model object
}
