% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit_model}
\alias{fit_model}
\title{Fits a \emph{pffr()} model to spatial summary function data}
\usage{
fit_model(formula, data, spatial_covars = NULL, nthreads = 1)
}
\arguments{
\item{formula}{An object of class \emph{formula}.  The formula for the \emph{pffr()} model with special terms as in \emph{mgcv's} "bam"}

\item{data}{Data frame obtained from the output of \emph{funboot::preprocess_data()}}

\item{spatial_covars}{Vector containing the names of the covariate(s) to be treated as spatially-varying.  For example, c('var1','var2')}

\item{nthreads}{Number of threads to use in the model fit}
}
\value{
A \emph{pffr()} model object
}
\description{
Fits a \emph{pffr()} model to spatial summary function data
}
