% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Preprocesses data}
\usage{
fit_model(formula, data, spatial.covars = NULL)
}
\arguments{
\item{sumfun.data}{is}

\item{covars}{is a vector of model covariate names found in sumfun.data}
}
\value{
It returns res.wildBS
}
\description{
Preprocesses data
}
