% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{format.spatial.variable}
\alias{format.spatial.variable}
\title{format.spatial.variable takes a spatial variable in long format and puts it into matrix format for input into pffr()}
\usage{
\method{format}{spatial.variable}(data, spatial.variable, grid, id)
}
\arguments{
\item{sumfun.data}{is}

\item{covars}{is a vector of model covariate names found in sumfun.data}
}
\value{
It returns res.wildBS
}
\description{
format.spatial.variable takes a spatial variable in long format and puts it into matrix format for input into pffr()
}
