% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ftest.R
\name{Ftest}
\alias{Ftest}
\title{Preprocesses data}
\usage{
Ftest(
  formula.full,
  formula.red,
  image.id = "image_number",
  patient.id = "patient_id",
  data,
  spatial.covars = NULL,
  B = 1000,
  alpha = 0.05,
  re = NULL,
  seed = 456
)
}
\arguments{
\item{grid}{the grid that the CBs were evaluated over.  eg. 0:200}
}
\value{
It returns res.wildBS
}
\description{
Preprocesses data
}
