% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ftest.R
\name{Ftest}
\alias{Ftest}
\title{Conducts a wild bootstrap partial F test for two nested \emph{pffr()} models}
\usage{
Ftest(
  formula.full,
  formula.red,
  data,
  spatial.covars = NULL,
  B = 500,
  re = NULL,
  seed = NULL
)
}
\arguments{
\item{formula.full}{An object of class \emph{formula}.  The formula for the full model with special terms as in mgcv's GAM}

\item{formula.red}{An object of class \emph{formula}.  The formula for the reduced model with special terms as in mgcv's GAM}

\item{data}{Data frame obtained from the output of \emph{funboot::preprocess_data()}}

\item{spatial.covars}{Vector containing the names of the covariate(s) to be treated as spatially-varying.  For example, c('var1','var2')}

\item{B}{Number of bootstrap samples}

\item{re}{Vector containing names of variable(s) for which to fit a random intercept.  For example, c('patient_id')}

\item{seed}{Random seed to be used during the bootstrap resampling (optional).}
}
\value{
A list containing the p-value from the F test and the bootstrap distribution of M
}
\description{
Conducts a wild bootstrap partial F test for two nested \emph{pffr()} models
}
