% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_fixed_and_re}
\alias{get_fixed_and_re}
\title{Fits a \emph{pffr()} model to spatial summary function data}
\usage{
get_fixed_and_re(model, n_Im_total, grid, id)
}
\arguments{
\item{model}{An object of class \emph{formula}.  The formula for the \emph{pffr()} model with special terms as in \emph{mgcv's} BAM}

\item{n_Im_total}{Data frame obtained from the output of \emph{funboot::preprocess_data()}}

\item{grid}{Vector containing the names of the covariate(s) to be treated as spatially-varying.  For example, c('var1','var2')}

\item{id}{Vector containing the names of the covariate(s) to be treated as spatially-varying.  For example, c('var1','var2')}
}
\value{
A \emph{pffr()} model object
}
\description{
Fits a \emph{pffr()} model to spatial summary function data
}
