% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_images.R
\name{plot_images}
\alias{plot_images}
\title{Preprocesses data}
\usage{
plot_images(data, images.to.plot, cell.types.to.plot, pallete)
}
\arguments{
\item{grid}{the grid that the CBs were evaluated over.  eg. 0:200}
}
\value{
It returns res.wildBS
}
\description{
Preprocesses data
}
