% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wildBS_CB.R
\name{plot_wildBS_CB}
\alias{plot_wildBS_CB}
\title{Plots wild bootstrap confidence bands obtained from \emph{funboot::EY_CB()}}
\usage{
plot_wildBS_CB(CB_object)
}
\arguments{
\item{CB_object}{List obtained from the output of \emph{funboot::EY_CB()}}
}
\value{
A \emph{ggplot} object
}
\description{
Plots wild bootstrap confidence bands obtained from \emph{funboot::EY_CB()}
}
