% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wildBS_CB.R
\name{plot_wildBS_CB}
\alias{plot_wildBS_CB}
\title{Plots wild bootstrap confidence bands obtained from \emph{funboot::wildBS_CB(),} along with pointwise confidence intervals}
\usage{
plot_wildBS_CB(CB.object, alpha = 0.05)
}
\arguments{
\item{CB.object}{List obtained from the output of \emph{funboot::wildBS_CB()}}

\item{alpha}{Desired significance level for the pointwise confidence intervals.  For example, 0.05.  The pointwise confidence intervals are obtained using the asymptotic normality of the statistic \eqn{\frac{\hat{\beta}_{j}(r)}{SE(\hat{\beta}_{j}(r))}} as calculated from \emph{pffr()} output, and are not adjusted for multiplicity.}
}
\value{
A \emph{ggplot} object for each model coefficient function
}
\description{
Plots wild bootstrap confidence bands obtained from \emph{funboot::wildBS_CB(),} along with pointwise confidence intervals
}
