% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocesses raw data}
\usage{
preprocess_data(
  data,
  from_cell,
  to_cell,
  qc_cellcount_cutoff = 0,
  P = 50,
  perm_yn = FALSE,
  R = 200,
  inc = 1,
  image_dims,
  summary_function = "L",
  verbose = TRUE
)
}
\arguments{
\item{data}{Data frame containing the variables \strong{patient_id}, \strong{image_number}, \strong{cell_id}, \strong{cell_x}, \strong{cell_y}, \strong{cell_type}, and any covariates to be adjusted for, as in the example data.}

\item{from_cell}{The "from" cell type to be used in the calculation of the spatial summary function}

\item{to_cell}{The "to" cell type to be used in the calculation of the spatial summary function}

\item{qc_cellcount_cutoff}{If an image has \strong{qc_cellcount_cutoff} or less "to" cells, or \strong{qc_cellcount_cutoff} or less "from" cells, the image is excluded from the calculation of the spatial summary function.  (Analogous to SpaceANOVA's \strong{Hard_ths} parameter.)}

\item{P}{The number of permutations to use in the simulation envelope procedure, as long as \strong{perm_yn}=TRUE.}

\item{perm_yn}{If TRUE, the permuted mean of the spatial summary function is calculated using the permutation envelope procedure described in the \href{https://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1009900}{Wilson paper}.  The outcome variable to be used in \emph{pffr()} is then calculated by subtracting the value of the permuted mean from the value of the observed spatial summary function.  If \emph{perm_yn}=FALSE, the outcome variable is calculated by subtracting the summary function's expected value assuming complete spatial randomness from the observed spatial summary function.}

\item{R}{The maximum radius at which the observed spatial summary function should be evaluated.}

\item{inc}{The increment of units by which the spatial summary function should be calculated.  For example, if \strong{R}=200 and \strong{inc}=1, the function is evaluated on the grid 0:200.}

\item{image_dims}{Vector containing the dimensions of the images on which cell locations are recorded, structured as c(x_min, x_max, y_min, y_max).  All images must be of the same dimension.}

\item{summary_function}{The spatial summary function to be calculated.  Must be one of 'K', 'L', or 'g'.}

\item{verbose}{If TRUE, prints progress}
}
\value{
A data frame containing values of the observed and expected spatial summary function, as well as values of the covariates from the input dataset. L_pmean = permuted mean from the permutation envelope, outcome = L_expect - L_obs if \strong{perm_yn}=FALSE, or L_expect - L_pmean if \strong{perm_yn}=TRUE.
}
\description{
Preprocess raw data by: 1.) filtering out sparse images using \strong{qc_cellcount_cutoff}, 2.) calculating the desired spatial summary function, 3.) calculating the outcome variable to be used in \emph{pffr()}
}
