% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocesses data}
\usage{
preprocess_data(
  data,
  from.cell,
  to.cell,
  qc.cellcount.cutoff = 20,
  P = 50,
  perm.yn = F,
  R = 200,
  inc = 1,
  image.dims = c(0, 1200, 0, 1200),
  summary.function = "g",
  seed = 456
)
}
\arguments{
\item{data}{is}

\item{from.cell}{is}

\item{to.cell}{is}

\item{qc.cellcount.cutoff}{is}

\item{P}{is}

\item{perm.yn}{is}

\item{R}{is}

\item{inc}{is}

\item{image.dims}{is}

\item{summary.function}{is}

\item{seed}{only makes a difference if perm.yn is switched on}
}
\value{
It returns sumfun.data
}
\description{
Preprocesses data
}
