% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildBS_CB.R
\name{wildBS_CB}
\alias{wildBS_CB}
\title{Calculates wild bootstrap confidence bands for \emph{pffr()} model coefficients}
\usage{
wildBS_CB(
  formula,
  data,
  spatial_covars = NULL,
  target,
  form,
  covar_list = list(),
  func,
  B1,
  B2,
  alpha = 0.05,
  re = NULL,
  id,
  nthreads = 1,
  inner_SEs_from_package = FALSE
)
}
\arguments{
\item{formula}{An object of class \emph{formula}.  The formula for the \emph{pffr()} model with special terms as in \emph{mgcv's} GAM}

\item{data}{Data frame obtained from the output of \emph{funboot::preprocess_data()}}

\item{spatial_covars}{Vector containing the names of the covariate(s) to be treated as spatially-varying.  For example, c('var1','var2')}

\item{B1}{Number of outer bootstrap samples}

\item{B2}{Number of inner bootstrap samples}

\item{alpha}{Desired significance level for the confidence band. For example, 0.05}

\item{re}{Vector containing names of variable(s) for which to fit a random intercept.  For example, c('patient_id')}

\item{inner_SEs_from_package}{Whether or not to replace the inner bootstrap loop standard errors with \emph{pffr()}'s estimated standard errors of the coefficient function.  Should only be used if the target quantity for the confidence band is a single model coefficient and the data contain one image per patient.}
}
\value{
A list containing the lower and upper bounds of the confidence band, as well as the estimated model coefficients and standard errors from the \emph{pffr()} output and additional intermittent variables used in the calculation of the confidence band.
}
\description{
Calculates wild bootstrap confidence bands for \emph{pffr()} model coefficients
}
