% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildBS_CB.R
\name{wildBS_CB}
\alias{wildBS_CB}
\title{Calculates wild bootstrap confidence bands for \emph{pffr()} model coefficients}
\usage{
wildBS_CB(
  formula,
  data,
  spatial.covars = NULL,
  B = 500,
  alpha = 0.05,
  re = NULL,
  seed = NULL
)
}
\arguments{
\item{formula}{An object of class \emph{formula}.  The formula for the \emph{pffr()} model with special terms as in mgcv's GAM}

\item{data}{Data frame obtained from the output of \emph{funboot::preprocess_data()}}

\item{spatial.covars}{Vector containing the names of the covariate(s) to be treated as spatially-varying.  For example, c('var1','var2')}

\item{B}{Number of bootstrap samples}

\item{alpha}{Desired significance level for the confidence band. For example, 0.05}

\item{re}{Vector containing names of variable(s) for which to fit a random intercept.  For example, c('patient_id')}

\item{seed}{Random seed to be used during the bootstrap resampling (optional).}
}
\value{
A list containing the lower and upper bounds of the confidence band, as well as the estimated model coefficients and standard errors from the \emph{pffr()} output and additional intermittent variables used in the calculation of the confidence band.  'p.vals' are p-values for each model coefficient function yielded by the wild bootstrap test of the model coefficient against 0, and 'test.stats' are the associated test statistics.
}
\description{
Calculates wild bootstrap confidence bands for \emph{pffr()} model coefficients
}
