% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildBS_CB.R
\name{wildBS_CB}
\alias{wildBS_CB}
\title{Preprocesses data}
\usage{
wildBS_CB(
  formula,
  data,
  spatial.covars = NULL,
  B = 1000,
  alpha = 0.05,
  re = NULL,
  seed = 456
)
}
\arguments{
\item{sumfun.data}{is}
}
\value{
It returns res.wildBS
}
\description{
Preprocesses data
}
